refactor hapus data pelatihan

+++ app/Http/Controllers/PelatihanController.php
@@ -38,10 +38,13 @@ class PelatihanController extends Controller
     return redirect()->Route('pelatihanData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
   }
 
-  public function Delete($Id){
-    $Id = HCrypt::Decrypt($Id);
-    $Pelatihan = Pelatihan::findOrFail($Id);
-    $Pelatihan->delete();
-    return redirect()->Route('pelatihanData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
+  public function Hapus($Id=null,$Verify=null){
+    if ($Verify) {
+      $Id = HCrypt::Decrypt($Id);
+      $Pelatihan = Pelatihan::findOrFail($Id);
+      $Pelatihan->delete();
+      return redirect()->Route('pelatihanData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
+    }
+    return abort(404);
   }
 }
